requirements.txt
streamlit==1.35.0
openai==1.30.1
faiss-cpu==1.7.4
langchain==0.2.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
tqdm==4.66.4
